% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{BoundBeta}
\alias{BoundBeta}
\title{Bound samples generated by rbeta}
\usage{
BoundBeta(...)
}
\arguments{
\item{...}{rbeta parameters}
}
\value{
bounded beta random variable
}
\description{
Bound samples generated by rbeta
}

