% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_normal.R
\docType{class}
\name{Normal}
\alias{Normal}
\title{R6 class for Normal node.}
\format{An \code{\link{R6Class}} generator object}
\usage{
Normal
}
\description{
R6 class for Normal node.
}
\section{Fields}{

\describe{
\item{\code{dataIds:}}{data IDs}

\item{\code{tssb:}}{A TSSB object}

\item{\code{children:}}{descentant node objects}

\item{\code{parent:}}{parent node object}
}}
\keyword{Normal}
\keyword{class}
\keyword{node}

