% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_pyclone.R
\docType{class}
\name{PyClone}
\alias{PyClone}
\title{R6 class for pyclone node.}
\format{An \code{\link{R6Class}} generator object}
\usage{
PyClone
}
\description{
R6 class for pyclone node.
}
\section{Fields}{

\describe{
\item{\code{dataIds:}}{data IDs}

\item{\code{tssb:}}{A TSSB object}

\item{\code{children:}}{descentant node objects}

\item{\code{parent:}}{parent node object}
}}
\keyword{PyClone}
\keyword{class}
\keyword{node}

