% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{RunNormal}
\alias{RunNormal}
\title{Fit a tssb with normal node with data via mcmc}
\usage{
RunNormal(data, numOfMCMC = 1000, burnIn = 0, dims = ncol(data),
  tssb = NULL, ...)
}
\arguments{
\item{data}{a data matrix or list}

\item{numOfMCMC}{mcmc iterations}

\item{burnIn}{burn-in number}

\item{dims}{dimension of data}

\item{tssb}{a use supplied TssbMCMC object}

\item{...}{additional parameters to node class generator}
}
\value{
A list of traces
}
\description{
Fit a tssb with normal node with data via mcmc
}

