% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{SticksToEdges}
\alias{SticksToEdges}
\title{Convert beta random variables to edges in stick-breaking process}
\usage{
SticksToEdges(sticks)
}
\arguments{
\item{sticks}{A beta random variable vector}
}
\value{
A vector of edges
}
\description{
Convert beta random variables to edges in stick-breaking process
}

