% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{TrimZeros}
\alias{TrimZeros}
\title{Trim leading or trailing zeros in a vector}
\usage{
TrimZeros(x, trim = "fb")
}
\arguments{
\item{x}{A numerical vector}

\item{trim}{A string with ‘f’ representing trim from front and ‘b’ to
trim from back. Default is ‘fb’, trim zeros from both front and back of the
array.}
}
\value{
The result of trimming the input
}
\description{
Trim leading or trailing zeros in a vector
}

