% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tssb.R
\docType{class}
\name{TSSB}
\alias{TSSB}
\title{R6 class for TSSB. TSSB is the basic object of tree structured stick-breaking
process}
\format{An \code{\link{R6Class}} generator object}
\usage{
TSSB
}
\description{
R6 class for TSSB. TSSB is the basic object of tree structured stick-breaking
process
}
\keyword{data}

