% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tssb_inference_batch_vb.R
\docType{class}
\name{TssbBatchVB}
\alias{TssbBatchVB}
\title{R6 class for inference via Batch variational Bayes for TSSB.}
\format{An \code{\link{R6Class}} generator object}
\usage{
TssbBatchVB
}
\description{
R6 class for inference via Batch variational Bayes for TSSB.
}
\keyword{data}

