% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tssb.R
\docType{class}
\name{TssbMCMC}
\alias{TssbMCMC}
\title{R6 class for inference via MCMC for TSSB.}
\format{An \code{\link{R6Class}} generator object}
\usage{
TssbMCMC
}
\description{
R6 class for inference via MCMC for TSSB.
}
\keyword{data}

