% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tssb.R
\docType{class}
\name{TssbSVB}
\alias{TssbSVB}
\title{R6 class for inference via stochastic variational Bayes for TSSB.}
\format{An \code{\link{R6Class}} generator object}
\usage{
TssbSVB
}
\description{
R6 class for inference via stochastic variational Bayes for TSSB.
}
\keyword{data}

