% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{Node}
\alias{Node}
\title{R6 class for Node. Node is the basic object of each cluster in TSSB}
\format{An \code{\link{R6Class}} generator object}
\usage{
Node
}
\description{
R6 class for Node. Node is the basic object of each cluster in TSSB
}
\section{Fields}{

\describe{
\item{\code{dataIds:}}{data IDs}

\item{\code{tssb:}}{A TSSB object}

\item{\code{children:}}{descentant node objects}

\item{\code{parent:}}{parent node object}
}}
\keyword{data}

