% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{RunNormal}
\alias{RunNormal}
\title{Fit a tssb with normal node with data via mcmc}
\usage{
RunNormal(data, numOfMCMC, tssb = NULL)
}
\arguments{
\item{data}{a data matrix}

\item{numOfMCMC}{mcmc iterations}
}
\value{
A list of traces
}
\description{
Fit a tssb with normal node with data via mcmc
}

