% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{SliceSampler}
\alias{SliceSampler}
\title{Sample from a user supplied distribution with slice sampler}
\usage{
SliceSampler(initX, logprob, sigma = 1, stepOut = T, maxStepsOut = 1000,
  compwise = F, ...)
}
\arguments{
\item{initX}{initial parameter}

\item{logprob}{a user supplied log-distribution function}

\item{sigma}{step size for slice construction}

\item{maxStepsOut}{maximum step out number}

\item{compwise}{use component wise update or not}

\item{...}{additional args for logprob}

\item{setpOut}{use setp out procedure or not}
}
\value{
a sample of parameter
}
\description{
Sample from a user supplied distribution with slice sampler
}

