% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tssb.R
\docType{class}
\name{TSSB}
\alias{TSSB}
\title{TSSB is a R6 object of tree-structured stick breaking
process}
\format{An \code{\link{R6Class}} generator object}
\usage{
TSSB
}
\description{
TSSB is a R6 object of tree-structured stick breaking
process
}
\section{Fields}{

\describe{
\item{\code{data}}{}

\item{\code{dpAlpha}}{}

\item{\code{dpGamma}}{}

\item{\code{dpLambda}}{}

\item{\code{minDepth}}{}

\item{\code{maxDepth}}{}

\item{\code{root}}{}

\item{\code{assignments}}{}

\item{\code{numOfData}}{}

\item{\code{minDpAlpha}}{}

\item{\code{maxDpAlpha}}{}

\item{\code{minDpLambda}}{}

\item{\code{minDpLambda}}{}

\item{\code{minDpGamma}}{}

\item{\code{maxDpGamma}}{}
}}
\keyword{data}

